requirements.txt
django==3.2
dj-database-url==1.2.0
gunicorn==20.1.0
tzdata==2022.7
pillow~=9.4.0
psycopg2-binary==2.9.6
django-environ==0.9.0
djangorestframework-simplejwt~=5.2.2
django-rest-swagger~=2.2.0
django-cors-headers~=3.14.0
django-admin-interface~=0.24.2
asgiref==3.5.2
sqlparse==0.4.2
django-smart-selects==1.5.9
djangorestframework~=3.13.1
django-environ~=0.9.0
requests~=2.28.1
aiogram==2.22.1
aiohttp~=3.8.1
